refactor pagination interfaces
