+++ src/frontend/src/components/common/Pagination/helpers/enums.ts
