+++ src/frontend/src/components/common/Pagination/helpers/types.ts
@@ -1,0 +1,7 @@
+export type PaginationProps = {
+  currentPage: number;
+  totalEntries: number;
+  pageSize: number;
+  changePage: (page: number) => void;
+  changePageSize: (pageSize: number) => void;
+};
